feat(Task-8): FE App level error alerts
