+++ packages/react-cloudfront-app/src/components/pages/admin/PageProductImport/components/CSVFileImport.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
-import axios, { AxiosError } from 'axios';
+import axios from 'axios';
 import Typography from '@mui/material/Typography';
 import Box from '@mui/material/Box';
-import { Alert, AlertColor, Snackbar, SnackbarOrigin } from '@mui/material';
 import { useGetAuthorizationToken } from '~/queries/authorization';
 
 type CSVFileImportProps = {
@@ -10,20 +9,8 @@ type CSVFileImportProps = {
   title: string;
 };
 
-const ALERT_POSITION: SnackbarOrigin = {
-  vertical: 'top',
-  horizontal: 'center',
-};
-
-const FULL_WIDTH = { width: '100%' };
-
-type AlertDisplayOptions =
-  | { isDisplayed: false }
-  | { isDisplayed: true; message: string; severity: AlertColor | undefined };
-
 export default function CSVFileImport({ url, title }: CSVFileImportProps) {
   const [file, setFile] = React.useState<File>();
-  const [alertOptions, setAlertOptions] = React.useState<AlertDisplayOptions>({ isDisplayed: false });
   const authToken = useGetAuthorizationToken();
   const onFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const files = e.target.files;
@@ -40,59 +27,28 @@ export default function CSVFileImport({ url, title }: CSVFileImportProps) {
   const uploadFile = async () => {
     if (!file) return;
 
-    try {
-      const response = await axios({
-        method: 'GET',
-        url,
-        headers: {
-          // Comment "Authorization" to simulate 401
-          Authorization: `Basic ${authToken}`,
-        },
-        params: {
-          fileName: encodeURIComponent(file.name),
-        },
-      });
+    const response = await axios({
+      method: 'GET',
+      url,
+      headers: {
+        Authorization: `Basic ${authToken}`,
+      },
+      params: {
+        fileName: encodeURIComponent(file.name),
+      },
+    });
 
-      const { signedUrl } = response.data ?? {};
+    const { signedUrl } = response.data ?? {};
 
-      await fetch(signedUrl, {
-        method: 'PUT',
-        body: file,
-      });
-
-      setAlertOptions({
-        isDisplayed: true,
-        message: `The "${file.name}" is successfully uploaded`,
-        severity: 'success',
-      });
-      setFile(undefined);
-    } catch (error) {
-      setAlertOptions({
-        isDisplayed: true,
-        message: error instanceof AxiosError ? error.message : String(error),
-        severity: 'error',
-      });
-    }
+    await fetch(signedUrl, {
+      method: 'PUT',
+      body: file,
+    });
+    setFile(undefined);
   };
-
-  const handleAlertClose = React.useCallback(() => setAlertOptions({ isDisplayed: false }), []);
 
   return (
     <Box>
-      <Snackbar
-        anchorOrigin={ALERT_POSITION}
-        open={alertOptions.isDisplayed}
-        autoHideDuration={4000}
-        onClose={handleAlertClose}
-      >
-        <Alert
-          onClose={handleAlertClose}
-          severity={alertOptions.isDisplayed ? alertOptions.severity : undefined}
-          sx={FULL_WIDTH}
-        >
-          {alertOptions.isDisplayed ? alertOptions.message : undefined}
-        </Alert>
-      </Snackbar>
       <Typography variant="h6" gutterBottom>
         {title}
       </Typography>
